+++ python/tests/test_package.py
@@ -31,7 +31,7 @@ def test_is_compiled_extension():
 
 
 def test_version():
-    assert __version__ == "0.2.4"
+    assert __version__ == "0.2.5"
 
 
 def test_extension_so_naming():